test(core-transactions): split handler.test.ts into multiple files, add additional tests
