+++ __tests__/unit/core-transactions/handlers/__fixtures__/htlc-secrets.ts
@@ -1,0 +1,3 @@
+export const htlcSecretHex = "my secret that should be 32bytes";
+
+export const htlcSecretHashHex = "317fcab2563aa3a1549fab4a83f9ec073f07db8ea081f7eded87b6f322fe9a09";
