+++ __tests__/unit/core-transactions/handlers/__mocks__/block-repository.ts
@@ -1,0 +1,14 @@
+let mockBlock: any | null;
+
+export const setMockBlock = (block: any | null) => {
+    mockBlock = block
+};
+
+export const blockRepository = {
+    getDelegatesForgedBlocks: async () => {
+        return mockBlock ? [mockBlock] : []
+    },
+    getLastForgedBlocks: async () => {
+        return mockBlock ? [mockBlock] : []
+    }
+};
