+++ __tests__/unit/core-transactions/handlers/two/multi-payment.test.ts
@@ -1,0 +1,195 @@
+import "jest-extended";
+
+import passphrases from "@arkecosystem/core-test-framework/src/internal/passphrases.json";
+import { BuilderFactory } from "@arkecosystem/crypto/src/transactions";
+import { Contracts, Application } from "@arkecosystem/core-kernel";
+import { Crypto, Enums, Interfaces, Managers, Transactions, Utils, } from "@arkecosystem/crypto";
+import { FactoryBuilder, Factories } from "@arkecosystem/core-test-framework/src/factories";
+import { Generators } from "@arkecosystem/core-test-framework/src";
+import { Identifiers } from "@arkecosystem/core-kernel/src/ioc";
+import { StateStore } from "@arkecosystem/core-state/src/stores/state";
+import { TransactionHandler } from "@arkecosystem/core-transactions/src/handlers";
+import { TransactionHandlerRegistry } from "@arkecosystem/core-transactions/src/handlers/handler-registry";
+import { Wallets } from "@arkecosystem/core-state";
+import { configManager } from "@packages/crypto/src/managers";
+import {
+    InsufficientBalanceError,
+} from "@arkecosystem/core-transactions/src/errors";
+import { setMockTransaction } from "../__mocks__/transaction-repository";
+import {
+    buildMultiSignatureWallet,
+    buildRecipientWallet,
+    buildSecondSignatureWallet,
+    buildSenderWallet,
+    initApp,
+} from "../__support__/app";
+
+let app: Application;
+let senderWallet: Wallets.Wallet;
+let secondSignatureWallet: Wallets.Wallet;
+let multiSignatureWallet: Wallets.Wallet;
+let recipientWallet: Wallets.Wallet;
+let walletRepository: Contracts.State.WalletRepository;
+let factoryBuilder: FactoryBuilder;
+
+const mockLastBlockData: Partial<Interfaces.IBlockData> = { timestamp: Crypto.Slots.getTime() , height: 4 };
+
+const mockGetLastBlock = jest.fn();
+StateStore.prototype.getLastBlock = mockGetLastBlock;
+mockGetLastBlock.mockReturnValue( { data: mockLastBlockData } );
+
+beforeEach(() => {
+    const config = Generators.generateCryptoConfigRaw();
+    configManager.setConfig(config);
+    Managers.configManager.setConfig(config);
+
+    setMockTransaction(null);
+
+    app = initApp();
+
+    walletRepository = app.get<Wallets.WalletRepository>(Identifiers.WalletRepository);
+
+    factoryBuilder = new FactoryBuilder();
+    Factories.registerWalletFactory(factoryBuilder);
+    Factories.registerTransactionFactory(factoryBuilder);
+
+    senderWallet = buildSenderWallet(factoryBuilder);
+    secondSignatureWallet = buildSecondSignatureWallet(factoryBuilder);
+    multiSignatureWallet = buildMultiSignatureWallet();
+    recipientWallet = buildRecipientWallet(factoryBuilder);
+
+    walletRepository.reindex(senderWallet);
+    walletRepository.reindex(secondSignatureWallet);
+    walletRepository.reindex(multiSignatureWallet);
+    walletRepository.reindex(recipientWallet);
+});
+
+describe("MultiPaymentTransaction", () => {
+    let multiPaymentTransaction: Interfaces.ITransaction;
+    let secondSignatureMultiPaymentTransaction: Interfaces.ITransaction;
+    let multiSignatureMultiPaymentTransaction: Interfaces.ITransaction;
+    let handler: TransactionHandler;
+
+    beforeEach(async () => {
+        const transactionHandlerRegistry: TransactionHandlerRegistry = app.get<TransactionHandlerRegistry>(Identifiers.TransactionHandlerRegistry);
+        handler = transactionHandlerRegistry.getRegisteredHandlerByType(Transactions.InternalTransactionType.from(Enums.TransactionType.MultiPayment, Enums.TransactionTypeGroup.Core), 2);
+
+        multiPaymentTransaction = BuilderFactory.multiPayment()
+            .addPayment("ARYJmeYHSUTgbxaiqsgoPwf6M3CYukqdKN", "10")
+            .addPayment("AFyjB5jULQiYNsp37wwipCm9c7V1xEzTJD", "20")
+            .addPayment("AJwD3UJM7UESFnP1fsKYr4EX9Gc1EJNSqm", "30")
+            .addPayment("AUsi9ZcFkcwG7WMpRE121TR4HaTjnAP7qD", "40")
+            .addPayment("ARugw4i18i2pVnYZEMWKJj2mAnQQ97wuat", "50")
+            .nonce("1")
+            .sign(passphrases[0])
+            .build();
+
+        secondSignatureMultiPaymentTransaction = BuilderFactory.multiPayment()
+            .addPayment("ARYJmeYHSUTgbxaiqsgoPwf6M3CYukqdKN", "10")
+            .addPayment("AFyjB5jULQiYNsp37wwipCm9c7V1xEzTJD", "20")
+            .addPayment("AJwD3UJM7UESFnP1fsKYr4EX9Gc1EJNSqm", "30")
+            .addPayment("AUsi9ZcFkcwG7WMpRE121TR4HaTjnAP7qD", "40")
+            .addPayment("ARugw4i18i2pVnYZEMWKJj2mAnQQ97wuat", "50")
+            .nonce("1")
+            .sign(passphrases[1])
+            .secondSign(passphrases[2])
+            .build();
+
+        multiSignatureMultiPaymentTransaction = BuilderFactory.multiPayment()
+            .addPayment("ARYJmeYHSUTgbxaiqsgoPwf6M3CYukqdKN", "10")
+            .addPayment("AFyjB5jULQiYNsp37wwipCm9c7V1xEzTJD", "20")
+            .addPayment("AJwD3UJM7UESFnP1fsKYr4EX9Gc1EJNSqm", "30")
+            .addPayment("AUsi9ZcFkcwG7WMpRE121TR4HaTjnAP7qD", "40")
+            .addPayment("ARugw4i18i2pVnYZEMWKJj2mAnQQ97wuat", "50")
+            .nonce("1")
+            .senderPublicKey(multiSignatureWallet.publicKey!)
+            .multiSign(passphrases[0], 0)
+            .multiSign(passphrases[1], 1)
+            .multiSign(passphrases[2], 2)
+            .build();
+    });
+
+    describe("bootstrap", () => {
+        it("should resolve", async () => {
+            setMockTransaction(multiPaymentTransaction);
+            await expect(handler.bootstrap()).toResolve();
+        })
+    });
+
+    describe("throwIfCannotBeApplied", () => {
+        it("should not throw", async () => {
+            await expect(handler.throwIfCannotBeApplied(multiPaymentTransaction, senderWallet, walletRepository)).toResolve();
+        });
+
+        it("should not throw - second sign", async () => {
+            await expect(handler.throwIfCannotBeApplied(secondSignatureMultiPaymentTransaction, secondSignatureWallet, walletRepository)).toResolve();
+        });
+
+        it("should not throw - multi sign", async () => {
+            await expect(handler.throwIfCannotBeApplied(multiSignatureMultiPaymentTransaction, multiSignatureWallet, walletRepository)).toResolve();
+        });
+
+        it("should throw if wallet has insufficient funds", async () => {
+            senderWallet.balance = Utils.BigNumber.ZERO;
+            await expect(handler.throwIfCannotBeApplied(multiPaymentTransaction, senderWallet, walletRepository)).rejects.toThrow(
+                InsufficientBalanceError,
+            );
+        });
+
+        it("should throw if wallet has insufficient funds send all payouts", async () => {
+            senderWallet.balance = Utils.BigNumber.make(150); // short by the fee
+            await expect(handler.throwIfCannotBeApplied(multiPaymentTransaction, senderWallet, walletRepository)).rejects.toThrow(
+                InsufficientBalanceError,
+            );
+        });
+    });
+
+    describe("apply", () => {
+        it("should be ok", async () => {
+            const senderBalance = senderWallet.balance;
+            const totalPaymentsAmount = multiPaymentTransaction.data.asset!.payments!.reduce(
+                (prev, curr) => prev.plus(curr.amount),
+                Utils.BigNumber.ZERO,
+            );
+
+            await handler.apply(multiPaymentTransaction, walletRepository);
+
+            expect(senderWallet.balance).toEqual(
+                Utils.BigNumber.make(senderBalance)
+                    .minus(totalPaymentsAmount)
+                    .minus(multiPaymentTransaction.data.fee),
+            );
+
+            for (const { recipientId, amount } of multiPaymentTransaction.data.asset!.payments!) {
+                const paymentRecipientWallet = walletRepository.findByAddress(recipientId);
+                expect(paymentRecipientWallet.balance).toEqual(amount);
+            }
+        });
+    });
+
+    describe("revert", () => {
+        it("should be ok", async () => {
+            const senderBalance = senderWallet.balance;
+            senderWallet.nonce = Utils.BigNumber.make(1);
+
+            for (const { recipientId, amount } of multiPaymentTransaction.data.asset!.payments!) {
+                const paymentRecipientWallet = walletRepository.findByAddress(recipientId);
+                paymentRecipientWallet.balance = amount;
+            }
+            const totalPaymentsAmount = multiPaymentTransaction.data.asset!.payments!.reduce(
+                (prev, curr) => prev.plus(curr.amount),
+                Utils.BigNumber.ZERO,
+            );
+
+            await handler.revert(multiPaymentTransaction, walletRepository);
+            expect(senderWallet.balance).toEqual(
+                Utils.BigNumber.make(senderBalance)
+                    .plus(totalPaymentsAmount)
+                    .plus(multiPaymentTransaction.data.fee),
+            );
+
+            expect(senderWallet.nonce.isZero()).toBeTrue();
+            expect(recipientWallet.balance).toEqual(Utils.BigNumber.ZERO);
+        });
+    });
+});
